--- packages/deposit/src/config/files/processing.ts
@@ -1,16 +1,0 @@
-const processing = [
-  {
-    label: "Test 1",
-    value: "t1",
-  },
-  {
-    label: "Test 2",
-    value: "t2",
-  },
-  {
-    label: "Test 3",
-    value: "t3",
-  },
-];
-
-export default processing;
