+++ packages/deposit/src/features/files/FilesTable.tsx
@@ -18,10 +18,9 @@ import ErrorRoundedIcon from '@mui/icons-material/ErrorRounded';
 import CircularProgress from '@mui/material/CircularProgress';
 import Tooltip from '@mui/material/Tooltip';
 import { useTranslation } from 'react-i18next';
-import { useAppSelector, useAppDispatch } from '../../app/hooks';
+import { useAppSelector, useAppDispatch } from '../../redux/hooks';
 import { getFiles, removeFile, setFileMeta } from './filesSlice';
-import fileRoles from '../../config/files/roles';
-import fileProcessing from '../../config/files/processing';
+import { fileRoles, fileProcessing } from './filesOptions';
 import type { SelectedFile, FileActionOptionsProps, FileItemProps } from '../../types/Files';
 import { dansUtilityApi, useCheckTypeQuery } from './api/dansUtility';
 import { LightTooltip } from '../generic/Tooltip';
