+++ packages/deposit/src/features/files/filesSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../../app/store';
+import { RootState } from '../../redux/store';
 import { SelectedFile, ReduxFileActions } from '../../types/Files';
 
 const initialState: SelectedFile[] = []
