+++ packages/deposit/src/features/metadata/fields/AutocompleteAPIField.tsx
@@ -13,11 +13,12 @@ import { useFetchGettyAATTermsQuery } from '../api/getty';
 import { useFetchSheetsQuery } from '../api/sheets';
 import { useFetchDatastationsTermQuery } from '../api/datastations';
 import { useFetchDansFormatsQuery } from '../../files/api/dansFormats';
-import { useAppDispatch, useAppSelector } from '../../../app/hooks';
+import { useAppDispatch, useAppSelector } from '../../../redux/hooks';
 import { getFieldStatus } from '../metadataHelpers';
 import { StatusIcon } from '../../generic/Icons';
 import { setField, setMultiApiField } from '../metadataSlice';
-import type { AutocompleteFieldProps, AutocompleteAPIFieldProps, TypeaheadAPI, ApiLinkProps, OptionsType } from '../../../types/Metadata';
+import type { AutocompleteFieldProps, AutocompleteAPIFieldProps, ApiLinkProps } from '../../../types/MetadataProps';
+import type { TypeaheadAPI, OptionsType } from '../../../types/MetadataFields';
 import type { QueryReturnType } from '../../../types/Api';
 import { lookupLanguageString } from '@dans-framework/utils/language';
 import Select from '@mui/material/Select';
