+++ packages/deposit/src/features/submit/submitSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../../app/store';
+import { RootState } from '../../redux/store';
 import type { 
   ReduxFileSubmitActions, 
   SubmittedFile,
