--- packages/deposit/src/uuid.d.ts
@@ -1,2 +1,0 @@
-// uuid declaration
-declare module 'uuid';
